backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "fastmcp>=2.3.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.60",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "structlog>=24.4.0",
    "uvicorn[standard]>=0.30.0",
]
